add initial implementation of username reminder and password reset message sending. password generation itself is stubbed out; i need to look for an algorithm and java implementation. also the email messages need to be i18n'd, a final format chosen (text/plain? text/html? multipart/alternative with both formats?), and the from address configured.
integrate javamail into the build and distro. the javamail session is
configured in the app server as a jndi resource. assume a plaintext smtp
server running on localhost, but provide config properties that let the admin
turn on starttls or use smtps instead.

add getUserByEmail method to ProvisioningManager and UserDAO and change
getUser(String username) to getUserByUsername.

change the stracktrace tag file handler to allow the exception to be a
Throwable instead of requiring it to be an Exception.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@292 5fd0b11a-f2f7-0310-948e-c8f80ebed640