set all event properties and parameters when creating an event. ensure that property and parameter nodes are only created when the corresponding icalendar property or parameter is actually set. add methods to set nodes for icalendar parameters. store tzid parameters.
all icalendar property nodes now have a jcr property for the icalendar
property value. this lets us store the raw value string so that we can use
that when exporting an icalendar view of the property (some values are
converted to multiple jcr properties and/or java native types which, when
recombined into a value string, could potentially differ from the original
input).


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@434 5fd0b11a-f2f7-0310-948e-c8f80ebed640