bug 5014: - instead of indexing virtual jcr properties when storing calendar data, use the same flattening technique to turn the hierarchical icalendar components/properties/parameters into a flattened list of jcr properties that are set on the calendar resource node.
with this change, report queries are now run against the calendar
resource nodes themselves, not against their jcr:content properties
(which is where the TextCalendarTextFilter indexed them).

the flattening logic for now has been copied from
TextCalendarTextFilter into JcrCalendarFlattener. TextFilterListener
no longer loads TextCalendarTextFilter. eventually we should remove
this listener altogether and incorporate XPathTimeRangeQueryBuilder
directly into our copy of jackrabbit.

for now we can revert back to the old behavior by changing the value
of CosmoConstants.INDEX_VIRTUAL_PROPERTIES to "true" and
recompiling. i will remove this capability once get some more testing
and verify that all of the reports tests work. i am especially
skeptical of the timerange and limitexpand tests - the results i got
with them didn't seem to match the expected results as listed in the
tests' readme.

also add the capability to JcrResourceMapper to copy a resource's
properties to a node (creating the node if necessary). this only works
for collections and calendar collections at the moment. over time we
should complete this capability for all model resource types and
convert the io subsystem and JcrTestHelper to using these classes.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@1443 5fd0b11a-f2f7-0310-948e-c8f80ebed640