bug 5176: fix NPE when processing DELTICKET without basic credentials
in order to delete a ticket, the authenticated principal must meet
one of these conditions:

1) admin user
2) owner of the ticket
3) a write ticket on this resource

note that there is a known issue when a DELTICKET fails because none
of these conditions are met. cosmo should return a 403 forbidden
response, but some layer above cosmo (perhaps springmodules, acegi
security, or tomcat) rewrites the response status to 501 not
implemented. i have opened this issue as bug 5340.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@1554 5fd0b11a-f2f7-0310-948e-c8f80ebed640