Default Cosmo view resolver.
This view resolver that forwards all views without
a "redirect:" or "forward:" prefix to a resource not found page.

This is done because all legal views should be resolved by a ResourceBundleViewResolver 
for i18n support, and to support nice resource not found messages.

Without this view resolver forwarding to an error page, control
would be returned to the ServletDispatcher and 404 pages would need
to be set in web.xml, which is undesirable for Cosmo.

git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@2790 5fd0b11a-f2f7-0310-948e-c8f80ebed640