User modification dialog dojo widget.
Depending on "createUser" parameter and username, this widget will use cmp to signup a user,
create a new user, modify a regular user's own account, or modify an arbitrary user's account.

Currently, this kind of overloads this widget, which may not be ideal. A better solution may
be to subclass a common ancestor widget. In any case, this works well for now.

git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@2802 5fd0b11a-f2f7-0310-948e-c8f80ebed640