bug: 7528
Update UserDao.createUser() to check for duplicate username ignoring case.
This will result in a DuplicateUsernameException being thrown for "test2"
and "TesT2".  The username is still stored without modifying the case.

Also update UserDao.updateUser() to make the same checks.  

Update UserDao to always store email as lowercase, to avoid two emails
being stored that are the same (email addresses are typically not case sensitive).


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@3149 5fd0b11a-f2f7-0310-948e-c8f80ebed640