bug: 6896
Add cascade delete to calendar property and timerange indexes.
This should speed up collection and item delete times a little as
the database now handles the deletion of the indexes when an item
is deleted.  Before, a SQL DELETE statement was issued for each
index.  There is still more work to be done to improve collection 
delete times, but this is a start.

git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@3232 5fd0b11a-f2f7-0310-948e-c8f80ebed640