bug 7573 - cosmo.version.txt now contains the proper version info
The fix was to work-around a maven 2 assembly bug by creating an empty cosmo.properties file
as it seems that even with filtering turned on the plugin doesn't really *do* any filtering
unless it sees a filter-properties file.  Even if it doesn't contain any values and is empty.

go figure.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@3333 5fd0b11a-f2f7-0310-948e-c8f80ebed640