Bug 7681: Include authenticated user in access logs.
New filter to put Cosmo authentication information into a request attribute.

Since it looks like the Acegi "container integration" is far more heavyweight than we need, 
this solution for getting Cosmo authentication information to the AccessLogValve is,
I believe, preferable.

Also, changing the  default AccessLogValve pattern from "combined" to the 
equivalent pattern (as specified in the AccessLogValve docs) with %u changed to 
%{COSMO_PRINCIPAL}r to take advantage of the new filter.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@3385 5fd0b11a-f2f7-0310-948e-c8f80ebed640