In commit 3504 I changed the logout controller to redirect to the login page.
Unfortunately, this means we are no longer redirecting to login_redirect.jsp to 
clear the appropriate cookies.

Fortunately, the work that was done on redirect_login.jsp can now be done in login.jsp, since login.jsp
is never seen by logged-in users. This also ensures that any time users see the login screen 
the appropriate cookies are cleared. Finally, this speeds up logout, since we don't need to 
run the dojo boilerplate twice (once on the redirect screen and once on the login screen).

Removing redirect_login.jsp and adding cookie clearing code to login.jsp. Also, fixing references
to redirect_login.jsp in env.js and views.properties.

git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@3518 5fd0b11a-f2f7-0310-948e-c8f80ebed640