whoops, forgot to commit this with r4700. Same comment:
autotriaging now works. When a delta is applied to an item, autoTriage() gets called.

Also, implict triageStatus for occurrences works as well. This is like autoTriaging excpet that it 
doesn't get persisted, it gets calculated each time getTriageStatus() is called. 

Note, autotriaging DOES happen on NoteOccurrences for which there are modificaitons.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@4701 5fd0b11a-f2f7-0310-948e-c8f80ebed640