Add pointInTime parameter to NoteItemTriageStatusComparator.  The rank is now calculated as the proximity of one of the timestamps (either triageStatusRank, eventStartDate, or lastModificedDate) to the point in time.
Fix issue with using TreeSet to limit the results.  Using a TreeSet with
a Comparator that allows objects with the same rank in the same set doesn't
really work.  Well it works to put the items in, but not for removing or
determining if the set contains a specific item.

git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@4837 5fd0b11a-f2f7-0310-948e-c8f80ebed640