introduce the notion of "service events". the idea is that listeners can register with services to receive events when service operations occur.
there are a couple different "scopes" of listener: request-scoped, which acts
on an event based on information provided in a protocol request and are
actually passed to the service method in its parameters, and server-scoped,
which are registered once at server startup and listen for every execution of
a service operation without regard to any particular request.

for now, only UserService.createUser fires an event, and it only operates on
request-scoped listeners. we can add wider event support as needed.

rework account activation to be implemented by an ActivationListener passed to
UserService.createUser by CmpServlet. OOTB collection and items are next.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@4847 5fd0b11a-f2f7-0310-948e-c8f80ebed640