Fix up timeout code.
To support this, I've created a new configuration system. Variables that should
be administrator configurable can now be added to src/main/webapp/webui.conf. This file will be
pulled from the server at runtime (and not included in the dojo build) and added to
cosmo.ui.conf.

UI timeout (the amount of time that can pass with no activity before a user is logged out) is now
determined as follows:

If cosmo.ui.conf.uiTimeout (as set in webui.conf) is set, we use that.
Otherwise, we look for the session timeout value, as set in WEB-INF/web.xml, or
the tomcat web.xml in the snarf distribution tomcat/conf directory.

Once the timeout has passed with no activity, the application's modal dialog pops up with a 
message telling the user they will be logged out in n seconds, where n is taken from webui.conf.
They may cancel the logout by clicking "cancel", or logout immediately by clicking "ok".

git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@4878 5fd0b11a-f2f7-0310-948e-c8f80ebed640