Dashboard DONE query optimization:
Only load maxDone items from the db that are explicitly
triaged DONE.  Pulling more than that is pointless and took
too long for large collections.  I had a collection where
700+ items were being loaded each DONE query.

Implement maxResults and orderBy in ItemFilter.  The orderBy
implementation is not as robust as I'd like, but theres not
much time left for 0.7 and a better implementation can be 
tackled in 0.8.  It turns out the ItemFilter model didn't 
quite fit the needs of the complicated dashboard queries.

git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@5192 5fd0b11a-f2f7-0310-948e-c8f80ebed640