More unicode problems.
py3k couldn't come sooner. I'm having a variety of problems with the underlying socket module, most likely because there is still a variety of ascii and unicode being sent.

I don't have time right not clean up all the code and make it 100% unicode, so I'm just going to skip usernames that can't be coerced to unicode.

-Mikeal



git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@5454 5fd0b11a-f2f7-0310-948e-c8f80ebed640