when returning a 403 response due to insufficient privileges, include DAV:needs-privileges precondition code in the response. when we're doing access control in the provider layer, we have enough information to also include the href and required privilege, but at the security filter layer we don't.
disallow MKTICKET and DELTICKET for ticket principals. it doesn't really
make any sense for a ticket to be the owning principal for another ticket.

allow tickets with DAV:read-free-busy but without DAV:read to PROPFIND
both DAV:current-user-privilege-set and ticket:ticketdiscovery.

refactor PROPFIND and REPORT access control checks to make more sense,
short-circuiting when the principal has DAV:read. return 404 rather than 403
when the principal does not have DAV:read for a PROPFIND and is not
looking for one of the two unprotected properties mentioned above.

add DavPrivilegeSet class that can go to and from xml and to and from a
ticket.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@5851 5fd0b11a-f2f7-0310-948e-c8f80ebed640