Server side work and unit tests for bug 8720.
Clients may now include a "subscription" element in the user representation at signup.

The body of the subscription element should be the uid of the collection to subscribe to. The subscription element must have two attributes "name" and "ticket" which should be the name and ticket key of the subscription.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@6076 5fd0b11a-f2f7-0310-948e-c8f80ebed640