bug 11418 When creating collection using ItemProvider, initialize the collections name to uuid.  This prevents items with the same name, which screws up dav, because dav wouldn't be able to figure out which item is addressed.
Do not copy over name in Item.copy().  There are instances where this results
in duplicate names in the same collection (for instance breaking a recurring event).
By not copying the name, the name of the copied item will be initialized with the
uuid of the item, resulting in unique names.

git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@6296 5fd0b11a-f2f7-0310-948e-c8f80ebed640