Move the rest of the images and css into the dojo-esque theme directory.
Also, pretty heavy modification of imagegrid code. We're no longer doing javascript-fu to set style properties based on the json file.
Instead, a new script has been introduced to transform the json file into a css file, which has been added to vc. Since the css really should
always be dynamically generated we might want to make this part of the build at some point, but given the infrequency with which the
image grid gets updated this probably won't be necessary for a little while.

This change has made the ui much snappier.


git-svn-id: http://svn.osafoundation.org/server/cosmo/trunk@6931 5fd0b11a-f2f7-0310-948e-c8f80ebed640